+++ mimeo/run_map.py
@@ -19,8 +19,9 @@ def mainArgs():
 	parser.add_argument('--gffout',type=str,default="mimeo_B_in_A.gff3",help='Name of GFF3 annotation file.')
 	parser.add_argument('--outfile',type=str,default="mimeo_alignment.tab",help='Name of alignment result file.')
 	parser.add_argument('--verbose',action="store_true",default=False,help='If set report LASTZ progress.')
-	parser.add_argument('--label',type=str,default="BHits",help='Set annotation TYPE field in gff.')
+	parser.add_argument('--label',type=str,default="BHit",help='Set annotation TYPE field in gff.')
 	parser.add_argument('--prefix',type=str,default="BHit",help='ID prefix for B-genome hits annotated in A-genome.')
+	parser.add_argument('--keeptemp',action="store_true",default=False,help='If set do not remove temp files.')
 	# LASTZ options
 	parser.add_argument('--lzpath',type=str,default="lastz",help='Custom path to LASTZ executable if not in $PATH.')
 	parser.add_argument('--minIdt',type=int,default=60,help='Minimum alignment identity to report.')
@@ -55,13 +56,13 @@ def main():
 		# Compose alignment commands
 		cmds = mimeo.map_LZ_cmds(lzpath=args.lzpath,pairs=pairs,minIdt=args.minIdt,minLen=args.minLen,hspthresh=args.hspthresh,outfile=outtab,verbose=args.verbose)
 		# Run alignments
-		mimeo.run_cmd(cmds,verbose=args.verbose)
+		mimeo.run_cmd(cmds,verbose=args.verbose,keeptemp=args.keeptemp)
 	#Import alignment as df
 	alignments = mimeo.import_Align(infile=outtab,prefix=args.prefix,minLen=100,minIdt=95)
 	# Write to GFF3
 	with open(gffout, 'w') as f:
 		for x in mimeo.writeGFFlines(alnDF=alignments,chrlens=chrLens,ftype=args.label):
 			f.write(x)
-	if tempdir and os.path.isdir(tempdir):
+	if tempdir and os.path.isdir(tempdir) and not args.keeptemp:
 		shutil.rmtree(tempdir)
 	print("Finished!")
